+++ lib/src/main/java/network/grape/lib/transport/TransportHeader.java
@@ -4,8 +4,8 @@ package network.grape.lib.transport;
  * Abstraction of common functions to transport layers
  */
 public interface TransportHeader {
-  short getSourcePort();
-  short getDestinationPort();
+  int getSourcePort();
+  int getDestinationPort();
 
   byte UDP_PROTOCOL = 6;
   byte TCP_PROTOCOL = 17;
